gcs_chat_app/requirements.txt
streamlit
PyPDF2
langchain
psycopg2-binary
numpy
google-cloud-aiplatform>=1.38.0
pgvector
langchain-community
langchain-google-vertexai
unstructured==0.7.12 # this version fixed
nltk
rapidocr-onnxruntime

